bump to version 1.0.2

+++ electron.js
@@ -1,4 +1,4 @@
-const {app, BrowserWindow, Menu, ipcMain} = require('electron');
+const {app, BrowserWindow, Menu, ipcMain, Tray} = require('electron');
 const {spawn} = require('child_process');
 const path = require('path');
 const isDev = process.env.NODE_ENV === 'development';
@@ -8,6 +8,7 @@ const store = new Store();
 let mainWindow;
 let serverProcess;
 let clientProcess;
+let tray;
 
 function createWindow() {
 
@@ -41,6 +42,13 @@ function createWindow() {
         mainWindow.show();
     });
 
+    mainWindow.on('close', (event) => {
+        if (!app.isQuitting) {
+            event.preventDefault();
+            mainWindow.hide();
+        }
+    });
+
     mainWindow.on('closed', () => {
         mainWindow = null;
     });
@@ -53,7 +61,16 @@ function startDevelopmentServers() {
         stdio: 'inherit'
     });
 
-    mainWindow.loadFile(path.join(__dirname, 'client/build/index.html'));
+    // Start React development server
+    clientProcess = spawn('npm', ['run', 'client'], {
+        cwd: __dirname,
+        stdio: 'inherit'
+    });
+
+    // Wait for servers to start, then load the React app
+    setTimeout(() => {
+        mainWindow.loadURL('http://localhost:3000');
+    }, 5000);
 
 }
 
@@ -71,7 +88,103 @@ function startProductionServer() {
     mainWindow.loadFile(path.join(__dirname, 'client/build/index.html'));
 }
 
-app.whenReady().then(createWindow);
+function createTray() {
+    const iconPath = path.join(__dirname, 'client/public/assets/icon.png');
+    tray = new Tray(iconPath);
+
+    const contextMenu = Menu.buildFromTemplate([
+        {
+            label: 'Show App',
+            click: () => {
+                if (mainWindow) {
+                    mainWindow.show();
+                    mainWindow.focus();
+                }
+            }
+        },
+        { type: 'separator' },
+        {
+            label: 'Play/Pause',
+            click: () => {
+                if (mainWindow) {
+                    mainWindow.webContents.send('tray-play-pause');
+                }
+            }
+        },
+        {
+            label: 'Playback Speed',
+            submenu: [
+                {
+                    label: '0.5x',
+                    click: () => {
+                        if (mainWindow) {
+                            mainWindow.webContents.send('tray-set-speed', 0.5);
+                        }
+                    }
+                },
+                {
+                    label: '1.0x',
+                    click: () => {
+                        if (mainWindow) {
+                            mainWindow.webContents.send('tray-set-speed', 1.0);
+                        }
+                    }
+                },
+                {
+                    label: '1.25x',
+                    click: () => {
+                        if (mainWindow) {
+                            mainWindow.webContents.send('tray-set-speed', 1.25);
+                        }
+                    }
+                },
+                {
+                    label: '1.75x',
+                    click: () => {
+                        if (mainWindow) {
+                            mainWindow.webContents.send('tray-set-speed', 1.75);
+                        }
+                    }
+                },
+                {
+                    label: '2.0x',
+                    click: () => {
+                        if (mainWindow) {
+                            mainWindow.webContents.send('tray-set-speed', 2.0);
+                        }
+                    }
+                }
+            ]
+        },
+        { type: 'separator' },
+        {
+            label: 'Quit',
+            click: () => {
+                app.isQuitting = true;
+                app.quit();
+            }
+        }
+    ]);
+
+    tray.setContextMenu(contextMenu);
+    tray.setToolTip('Storytel Player');
+
+    tray.on('double-click', () => {
+        if (mainWindow) {
+            if (mainWindow.isVisible()) {
+                mainWindow.hide();
+            } else {
+                mainWindow.show();
+                mainWindow.focus();
+            }
+        }
+    });
+}
+
+app.whenReady().then(() => {
+    createWindow();
+    createTray();
+});
 
 ipcMain.handle('store-get', (event, key) =>  store.get(key));
 ipcMain.handle('store-set', (event, key, value) => store.set(key, value));
@@ -79,16 +192,20 @@ ipcMain.handle('store-remove', (event, key) => store.delete(key));
 
 
 app.on('window-all-closed', () => {
-    // Kill server processes
-    if (serverProcess) {
-        serverProcess.kill();
-    }
-    if (clientProcess) {
-        clientProcess.kill();
-    }
+    // Don't quit the app when all windows are closed, just hide to tray
+    // Only quit when explicitly requested through tray menu
+    if (app.isQuitting) {
+        // Kill server processes
+        if (serverProcess) {
+            serverProcess.kill();
+        }
+        if (clientProcess) {
+            clientProcess.kill();
+        }
 
-    if (process.platform !== 'darwin') {
-        app.quit();
+        if (process.platform !== 'darwin') {
+            app.quit();
+        }
     }
 });
